backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
recipe-scrapers==14.55.0
pillow==10.2.0
pytesseract==0.3.10
anthropic==0.18.1
openai==1.12.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0
authlib==1.3.0
itsdangerous==2.1.2
starlette-session==0.4.3
cryptography==42.0.0
slowapi==0.1.9
